refactor: move to separate component
