+++ app/src/main/kotlin/com/aliucord/manager/MainActivity.kt
@@ -5,23 +5,18 @@
 
 package com.aliucord.manager
 
-import android.content.Intent
-import android.os.*
-import android.provider.Settings
+import android.os.Bundle
+import android.os.Environment
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.*
-import androidx.activity.result.contract.ActivityResultContracts
+import androidx.activity.compose.BackHandler
+import androidx.activity.compose.setContent
 import androidx.compose.animation.*
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.window.DialogProperties
-import androidx.core.net.toUri
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.aliucord.manager.domain.manager.PreferencesManager
+import com.aliucord.manager.ui.dialog.ManageStorageDialog
 import com.aliucord.manager.ui.navigation.AppDestination
 import com.aliucord.manager.ui.screen.*
 import com.aliucord.manager.ui.theme.ManagerTheme
@@ -47,39 +42,12 @@ class MainActivity : ComponentActivity() {
                 isDynamicColor = preferences.dynamicColor
             ) {
                 val navigator = rememberBackstackNavigator<AppDestination>(AppDestination.Home)
-                var manageStorageGranted by remember { mutableStateOf(Environment.isExternalStorageManager()) }
 
                 BackHandler {
                     if (!navigator.pop()) finish()
                 }
 
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R && !manageStorageGranted) {
-                    AlertDialog(
-                        onDismissRequest = {},
-                        confirmButton = {
-                            val launcher = rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) {
-                                if (Environment.isExternalStorageManager())
-                                    manageStorageGranted = true
-                            }
-
-                            Button(
-                                onClick = {
-                                    Intent(Settings.ACTION_MANAGE_APP_ALL_FILES_ACCESS_PERMISSION)
-                                        .setData("package:com.aliucord.manager".toUri())
-                                        .let { launcher.launch(it) }
-                                },
-                            ) {
-                                Text(stringResource(R.string.open_settings))
-                            }
-                        },
-                        title = { Text(stringResource(R.string.permissions_grant)) },
-                        text = { Text(stringResource(R.string.permissions_grant_body)) },
-                        properties = DialogProperties(
-                            dismissOnBackPress = false,
-                            dismissOnClickOutside = false
-                        )
-                    )
-                }
+                ManageStorageDialog()
 
                 Taxi(
                     modifier = Modifier.fillMaxSize(),
